requirements.txt
streamlit==1.50.0
sentence-transformers==5.1.1
qdrant-client==1.15.1
groq==0.9.0
requests==2.32.3
httpx==0.28.1
python-dotenv==1.0.1
numpy==1.26.4
faiss-cpu==1.8.0.post1
